fix des résistances, tests et continuation du joueur

+++ src/resistance.h
@@ -7,14 +7,14 @@
 #include "graph.h"
 #include "graph_aux.h"
 
-float get_resistance(const struct graph_t* g, int color, int i, int j);
+double get_resistance(const struct graph_t* g, int color, int i, int j);
 
-float** generate_meshes(const struct graph_t* g, int color);
+double** generate_meshes(const struct graph_t* g, int color);
 
-void free_sys(float** mat, int n);
+void free_sys(double** mat, int n);
 
-void gauss(float** mat, float* b, float* x, int n);
+void gauss(double** mat, double* b, double* x, int n);
 
-float get_ratio(const struct graph_t* g, struct move_t mec);
+double get_ratio(const struct graph_t* g, struct move_t mec);
 
 #endif
